Reworked Compactors
Changed the API of the compactors to provide only one `compact()` method
with both the file name and the file contents. There is two reasons for
this change:

- The API results in greater cohesion: there is no need to think of a
  support usage first, it's no longer possible to call compact on an
unsupported file by accident.
- When adding a PhpScoper compactor, it will need both the file name and
  file content. It would be possible to do it with the current API by
storing the file as a property to be able to access to it in the compact
method but that would be relatively awkward and fragile.