Refactor configuration to leverage utility classes (#34)
Paves the way for #32.

As it stands, #32 requires some functions from `Configuration` which would require `Configuration` to be serializable. This is a lot of work so it makes sense to rather encapsulate some behaviour in utility classes and leverage them instead. As a result only those utility classes would require to be serializable which is trivial.

This also prepares a change for the `Box` class which IMO should handle the file mapping as it already handles the placeholders replacement and compactor processing.