Aggregate files
Nothing changes configuration-wise for the user, but the Configuration
class is drastically simplified by aggregating all the files (files,
directories, finders accounting for the blacklist) and doing the same
for the binary files.

When building a PHAR in very verbose mode, all the files where listed.
This is no longer done as will no longer be doable in the future (at
least easily) when refactoring the Box class to build from an iterator
rather than handling a file by file case in the Build command.