Fix error handling
Instead of calling `set_error_handler()` in the Application constructor,
the error handler is registered in `bin/box`. This allows to avoid some
pains from using the application with PHPUnit as PHPUnit registers its
own error handler.

Remove as well the usage of silenced error reporting combined with
exceptions get_last_error pattern in favour of:

- Better check: for example check that a file is writeable before
  attempting to write in it
- Letting the error be thrown directly (thanks to the error handler)

As a result, the custom exceptions have been removed.