Dump files instead of adding strings
Although this patch contains a couple of other things, the main change
here is how the files are added.
Indeed right now the current process is to process the contents of the
files in parallel and then add them one by one to the PHAR.
Instead, this new approach will dump all of those files in a temporary
location to take advantage of the `Phar::buildFromDirectory()` method
which seems to be order of magntitudes more performant.

Other changes:

- Add a missing rule to PHP-CS-Fixer to get rid of unused use statements
- Adjust PHP-CS-Fixer list config to use the short syntax instead of the
  long one
- Exclude the tests classes (not test cases) from the non-dev code
- Create a new TestCase class to deal with temporary directories
- Register the main script before the files instead of after
- Tweak temporary dir creation to do only 10 attemps maximum