Fix canonicalize (#61)
In some cases it was more convenient to get the directory path with an ending slash e.g. `dir/`. To achieve that the `canonicalize` function was "hacked", i.e. instead of being a simple proxy to `Path::canonicalize()` it was doing some extra stuff to keep that ending slash.

However this make the `canonicalize` function incorrect. Instead the issue has been fixed at its core: `canonicalize` behaves in a more "standard" way and cases where the ending slash is required are fixed.