Fix usage of symlinks (#74)
Using symlinks in `files{-bin}` and `directories{-bin}` was causing an `InvalidArgumentException` error stating that the file/directory could not be found. When used with the Finder, the files were added but later when adding them to Box an error was occurring as the file could not be found.

Instead a more human-friendly error is now thrown and errors will occurs when the configuration will be loaded (not later).