Include all files when no file setting is used (#75)
A recent feature is to be able to run Box without any configuration file (#73). In this scenario, all the files found are included. This PR allows the user to create the configuration file without losing this ability. Even better, he can use the `blacklist` setting to start to filter out some files. As soon as a file setting is used (even if empty) however, the files settings will be used to collect the files.