Dump all files in debug mode
Debug mode now dumps all the files (the binary files were currently not included). The handling of
the dump has been changed a bit as well: the previous dumped files are now removed in `Compile`
which is more natural and less hacky than in `Box::addFile()`.

The configuration file is dumped as well to ease debugging.

The verbosity is forced to debug when the debug mode is enabled.