Fix the main script inclusion
Since the main script requires some special processing and is added before the regular and binary
files, it should not be included in the regular or binary files.

This was actually the case when the main script was included via `files`, `files-bin` and `append`
(from `finder` or `finder-bin`). The main script path is now properly excluded in such cases.

Closes #168