Start Amp only if PHP-Scoper is used
Starting Amp has an overhead and increase the likelyhood of rising an issue due to how Amp works. As
a result, it is better to not have to start it unless necessary. So far the current compactors are
fast enough to not require parallel processing, the build from Box without PHP-Scoper goes from 5s
with parallel processing to 3s without.

Closes #135