Refactor the way files are added to the PHAR
No longer rely on the exposed PHAR by the Box class and instead introduce a new API. The new API is
meant to keep all the added files in memory and add them to the PHAR in one shot at the very end.
This will make it easier to decide whether or not the autoload should be dumped and no longer worry
about the order in which the files are added regarding the autoload.