Allow dumping requirements from the composer.json file (#203)
When there is no dependencies Composer does not generate a `composer.lock` and as a result the
requirement checker could not be generated. This patch now allows to use the requirement checker
even in that situation.

Note however that this implies the requirement checker could be wrong if one has a `composer.json`
with dependencies but the `composer.lock` file went missing for some reasons. This is however and
edge case that I don't think Box should worry about.