Fix the dump of the requirement checker
The issue is that some classes for `.requirement-checker/src` had a prefix but the ones from
`.requirement-checker/vendor`.

This issue went unnoticed in the CI because when running `make test` the `.requirement-checker` is
actually redumped (even though it should not).

Makefile being a bit silly here, `.requirement-checker` is now dumped by a make command and no
longer dumped during the tests