Fix the processing of files (#320)
When running the `process` command, the file path provided to the compactors was the absolute path whereas the `Box` class provides the relative path. This was causing issues with the PHP-Scoper patchers since one will likely expect the relative path and not account for the path to be absolute.
To avoid extra unintuitive work to the user, the relative path is provided as during a regular
compilation.

When running the `process` command with a `quiet` verbosity, only the processed file contents will
be outputed. This allows one to easily copy the processed file contents.