Fix the validate command
No longer make `Validate` extend `ConfigurableCommand` since the configuration file is an argument
and the argument should be used instead of the option.

The validate command now properly takes the argument given instead of relying on the default find
found in the current working directory or the file passed via the configuration option.