Make the requirement checker dump deterministic (#410)
The requirement checker needs to be dumped whenever Box itself or the requirement checker changes. However as of now, the prefix used is random causing the git status to be constantly polluted by a newly dumped requirement checker.

In this PR, the prefix is constant but defined by the last release version which will fix the problem described above whilst ensuring each Box release will have a different requirement checker shipped.