Compactor\Php: add support for PHP 8.0+ attributes
Up to now, when a Phar was being created using PHP < 8.0 with the PHP compactor enabled, PHP 8.0+ attributes would be incorrectly removed as if they were comments.

This has now been fixed.

Both the solution implemented, as well as the unit tests, are largely inspired by the solution for handling attributes when code is being tokenized on PHP < 8.0, as implemented in `PHP_CodeSniffer` in https://github.com/squizlabs/PHP_CodeSniffer/pull/3203 and https://github.com/squizlabs/PHP_CodeSniffer/pull/3299. Props to alekitto.

Fixes 567

Co-authored-by: Alessandro Chitolina <alekitto@gmail.com>