Detect if the Box PHAR is scoped (#632)
I found that the current Box PHAR is not scoped due to a patcher issue. There is unfortunately no way to fix the detection of a scoping failure within Box as this happens within the compactor (unlike in PhpScoper where there is an info log) which can be executed in a different process.

Still, by printing the current namespace, we can easily check the namespace is a PHPScoper prefix or not.