Composer: minor tweaks
While updating a workflow to use Box 4.x, I noticed that the branch alias hadn't been updated.

Might also be worth it to mention in the changelog that the `sodium` extension is now a non-dev requirement for running Box (under BC-break).
The entry as-it-is under `Misc` was the reason for to check the `composer.json` of the file to see what was required as an alternative, only to realize that the extension is now required.