ci: Fix running tests with PHAR readonly or writeable
The make commands `make phpunit_phar_readonly` and `make phpunit_phar_writeable` are helpful but not good enough for the CI. Indeed the tests executed in different processes are not inheriting the ini values set at runtime. So doing `php -dphar.readonly=0` will not work for the tests with `@runInSeparateProcesses`.