feat: Deprecate explicitely enabling the requirement checker when the composer.lock could not be found. (#867)
In some older versions of Composer the `composer.lock` file was not always generated. It is now, at the very least since 2.2, always generated.

Since the `composer.json` information is too lacking to provide a reliable result for the requirement checker (it does not contain the information about each package, hence what the packages requirements are or what extension they provide cannot be known), I want to deprecate using the requirement checker when the `composer.lock` file is not found.