fix: Take the polyfilled extensions from the "provide" package info
When we have access to a package info in a `composer.lock`, which
extension a package may polyfill can be found in the `provide` section.
This is much more reliable than trying to guess it from some package
conventions like the Symfony polyfills or manually map some other
popular polyfills.

The previous way of doing it kept for now, in case we fall back to the
`composer.json`, but it will disappear as soon as we no longer support
inferring the requirements from the `composer.json` file in 5.0.