refactor: Inline the pharaoh package (#895)
The [pharaoh package](https://github.com/paragonie/pharaoh) is unfortunately no longer maintained. I offered several patches but in the end most of it had to be patched directly in Box.

I think there is a high risk of being a liability in the future and this is also making it harder to improve the diff & info Box commands.

In order to bring improvements, I think it's best to "inline" this package for now. Once the feature set desired is reached and the API more stable, it will still be possible to extract it or port back those elements.