feat: Rework extract command (#961)
Dump extra information when extracting a PHAR. This is mostly the PHAR specific data that you can find (in different format sometimes) in the PHAR `.phar` directory. Dumping this information allows to get this information without keeping around a copy of the original PHAR and without using a `Phar` instance. This matters little for the extract command itself but that means later one we can get all the information of a PHAR by extracting it in a separate process and then not having to load a `Phar` instance anymore.