feat: No longer fail on non satisfiable PHP constraint when generating Docker file
When generating a Dockerfile the PHP image token is a "nice to have".
The docker image itself is meant as generating a helpful template rather
than a definite file that should not be touched.

With this in mind, failing to generate the template because the
appropriate PHP image could not be determine is rather annoying as it is
something easily fixable by the user whilst the list of templates is a
bit more of an inaccessible content.