fix: Fix the autoloading of the excluded files (#1323)
This fixes the integration of
https://github.com/humbug/php-scoper/pull/864 within Box which was done in #1142.

The issue was that the relative file paths was passed instead of hashes. Note that we create the file hasher in a different way than in PHP-Scoper, since PHP-Scoper has to deal with the vendor directory from the source and the scoped one, and deals with absolute paths. In Box however, this is not the case, we directly get relative file paths (relative to the working directory)