perf: Improve performance of Box::collectRemainingFiles()
The `->notPath()` filter of Symfony Finder is either accepting a “literal path”
or a regular expression and converts each literal path to a regular expression
and then filters the file stream by applying each regex individually. This
leads to O(bufferedFileNames^2) calls to `preg_match()`, well over 20 million
for PHPStan.

Fix this by performing a simple `in_array()` check on the buffered files. This
brings down the time to compile PHPStan from 2:35 to 0:25 on my machine.

I have verified that the resulting PHPStan Phar hash is identical before and
after this change.

Fixes box-project/box#1551.